requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
azure-core==1.36.0
azure-identity==1.25.1
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
fonttools==4.61.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
microsoft-kiota-abstractions==1.9.7
microsoft-kiota-authentication-azure==1.9.7
microsoft-kiota-http==1.9.7
microsoft-kiota-serialization-form==1.9.7
microsoft-kiota-serialization-json==1.9.7
microsoft-kiota-serialization-multipart==1.9.7
microsoft-kiota-serialization-text==1.9.7
msal==1.34.0
msal-extensions==1.3.1
msgraph-core==1.3.8
msgraph-sdk==1.49.0
multidict==6.7.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.2.1
requests==2.32.5
std-uritemplate==2.0.8
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
yarl==1.22.0
zipp==3.23.0
zopfli==0.4.0
